♻️  prefer consistent, grouped import paths (and sort them if possible)
Consistency avoids many pitfalls, and helps searchability. Always good.

Grouping and sorting imports is a great way to improve reviewability
debugging, and avoid stupid mistakes.

If there's bug, a name conflict, a symbol imported from the wrong
place, it's way easier to track what is going on when imports are tidy.
Moreover, it can help grasp at first glance what kind of content the
component is a hub for. It also tells about a component's complexity.

I did it manually here, but really it would be a good idea to
configure eslint with the import extension to do this automatically.

In javascript, groups are typically this:
1. external dependencies
2. js code
3. components and assets

As usual, writing code is way easier than reading it.
Auto-imports help importing stuff, but without proper configuration,
they make a mess, so it's a thing to keep in mind.